+++ src/client/components/ac_input_email/ac_input_email.js
@@ -104,7 +104,7 @@ Template.ac_input_email.helpers({
 
     // whether the mandatory field must exhibit an ad-hoc colored border ?
     mandatoryBorder(){
-        return Template.instance().AC.mandatoryField() && Template.currentData().component.opts().mandatoryFieldsBorder() ? 'ac-mandatory-border' : '';
+        return Template.instance().AC.mandatoryField() && Template.currentData().component.opts().coloredBorders() === AC_COLORED_MANDATORY ? 'ac-mandatory-border' : '';
     },
 
     // returns the translated string